+++ project_task_subtask/doc/changelog.rst
@@ -1,3 +1,10 @@
+
+`1.1.0`
+-------
+
+- **Improvement**: 'Checklist' replaced by 'To-Do List'
+- **New**: added group "Allow edit others tasks"
+
 `1.0.0`
 -------